% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bg.R
\name{add_bg}
\alias{add_bg}
\title{add_bg}
\usage{
add_bg(wb, sheet, rows, cols, color)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{sheet name}

\item{rows}{rows the style should be added to}

\item{cols}{columns the style should be added to}

\item{color}{fillcolor of the specified cells (Rcolor or HEX Value)}
}
\description{
Adds fill color to specified cells
}
