% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_spalte_zeile.R
\name{add_data_spalte_jahr}
\alias{add_data_spalte_jahr}
\title{Function to append a year as column}
\usage{
add_data_spalte_jahr(wb, year, data, dataStart, limit = NULL, sheet = 1)
}
\arguments{
\item{wb}{a workbook object}

\item{year}{the year of the data}

\item{data}{A dataset with one column where the name is the respective year. Can also be a list of datasets}

\item{dataStart}{row index where the data starts}

\item{limit}{limit for small values. Values below this threshold will be marked with "-"}

\item{sheet}{sheet number. default is 1. If more than one sheet, a range can be passed to the function}
}
\value{
workbook object
}
\description{
Appends a data column where the year has to be the name of the column
}
