% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_spalte_zeile.R
\name{add_data_zeile_jahr}
\alias{add_data_zeile_jahr}
\title{Function to append a year as row}
\usage{
add_data_zeile_jahr(wb, year, data, dataStart, sheet = 1)
}
\arguments{
\item{wb}{a workbook object}

\item{year}{the year of the data}

\item{data}{A dataset with one row where the column "jahr" contains the respective year. Can also be a list of datasets}

\item{dataStart}{row index where the data starts}

\item{sheet}{sheet number. default is 1. If more than one sheet, a range can be passed to the function}
}
\value{
workbook object
}
\description{
Appends a data row where the the year has to be in one variable "jahr"
}
