% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_datenquelle.R
\name{add_datenquelle}
\alias{add_datenquelle}
\title{add_datenquelle}
\usage{
add_datenquelle(wb, sheet, startRow, col = 1, datenquelle)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{startRow}{row where to start}

\item{col}{column where to insert}

\item{datenquelle}{}
}
\description{
Function to add datenwuelle
}
