% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_decoration.R
\name{add_decoration}
\alias{add_decoration}
\title{add_decoration}
\usage{
add_decoration(wb, sheet, rows, cols, decoration)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{sheet name}

\item{rows}{rows the style should be added to}

\item{cols}{columns the style should be added to}

\item{decoration}{see [openxlsx::createStyle()]}
}
\description{
Adds a font decoration to specified cells
}
