% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_footnotes.R
\name{add_footnotes}
\alias{add_footnotes}
\title{add_footnotes}
\usage{
add_footnotes(wb, sheet, startRow, col = 1, footnotes)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{startRow}{row where to start}

\item{col}{column where to insert}

\item{footnotes}{character vector or string with footnotes}
}
\description{
Function to add footnodes
}
