% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_special_numfmt.R
\name{add_special_numfmt}
\alias{add_special_numfmt}
\title{add_special_numfmt}
\usage{
add_special_numfmt(
  wb,
  sheet,
  data,
  pattern,
  startCol,
  dataStart,
  position_prefix = TRUE
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{name of the sheet}

\item{data}{a data.frame. should only conatin the numbers you want to write to the excel but no leading column with string information}

\item{pattern}{a regular expression that should be looked for in the data}

\item{startCol}{where the data insert should start}

\item{dataStart}{row in which the numeric data in the file starts (e.g. first row after header)}

\item{position_prefix}{if `TRUE` the pattern is inserted as prefix, otherwiese as suffix (default is prefix)}
}
\description{
Function writes data to workbook and adds special numFmt for matching cells
}
