% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_style_feature.R
\name{add_style_feature}
\alias{add_style_feature}
\title{add_style_feature}
\usage{
add_style_feature(wb, sheet, rows, cols, feature_names, feature_values)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{sheet name}

\item{rows}{rows the style should be added to}

\item{cols}{columns the style should be added to}

\item{feature_names}{a character vector including the names of the features you wish to change (see [openxlsx::createStyle()])}

\item{feature_values}{a vector of values for the  corresponding features from `feature_names`}
}
\description{
A Function to change specific parameters of a cellstyle without overwriting it.
}
