% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_data_column.R
\name{append_data_column}
\alias{append_data_column}
\title{append_data_column}
\usage{
append_data_column(wb, sheet, varname, data, dataStart)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{varname}{name of the new variable}

\item{data}{a data.frame with the data you want to insert}

\item{dataStart}{row where the data starts}
}
\description{
Append column to the end of the data
}
