% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_header_data.R
\name{copy_header_data}
\alias{copy_header_data}
\title{copy_header_data}
\usage{
copy_header_data(wb, sheet, newSheet, dataStartRow)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object from which to copy the data}

\item{newSheet}{new sheet where the data should be copied}

\item{dataStartRow}{row where the body data starts}
}
\description{
Copy the header without formating from one sheet to another.
}
