% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_heights_widths.R
\name{copy_heights_widths}
\alias{copy_heights_widths}
\title{copy_heights_widths}
\usage{
copy_heights_widths(wb, sheet, newSheet)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object from which to copy the row heights and col widths}

\item{newSheet}{new sheet where the row heights and col widths should be inserted}
}
\description{
Copy colWidths and rowHeights from one sheet to another one
}
