% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_sheet_formats.R
\name{copy_sheet_formats}
\alias{copy_sheet_formats}
\title{copy_sheet_formats}
\usage{
copy_sheet_formats(wb, sheet, newSheet)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object from which to copy the styles}

\item{newSheet}{new sheet where the styles should be copied}
}
\description{
Copy all styles from one sheet to another one
}
