% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_style.R
\name{create_data_style}
\alias{create_data_style}
\title{create_data_style}
\usage{
create_data_style(
  wb,
  sheet,
  startRow,
  date = NULL,
  year = NULL,
  startCol = 1,
  data,
  datenquelle = NULL,
  gemeinde = FALSE
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{startRow}{row in which the data should start}

\item{date}{index of a date variable (default=NULL)}

\item{year}{index of a year variable (default=NULL)}

\item{startCol}{col in which the data should start}

\item{data}{a data.frame}

\item{gemeinde}{if true formating for Gemeinde tables will be included (bold for Bezirk and Kanton) (default = FALSE)}
}
\description{
create_data_style
}
