% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_directory}
\alias{create_directory}
\title{Create directory from file path}
\usage{
create_directory(xlsx_path)
}
\arguments{
\item{xlsx_path}{file path}
}
\description{
Create directory from file path
}
