% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_header_style.R
\name{create_header_style}
\alias{create_header_style}
\title{create_header_style}
\usage{
create_header_style(
  wb,
  sheet,
  ncol,
  text,
  startRow = 1,
  style = header_bold_12_left,
  rowHeight = 18.75,
  ...
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{ncol}{the number of columns that the header should be long}

\item{text}{the text of the header}

\item{startRow}{row in which the header should be placed (default = 1)}

\item{style}{an openxlsx::Style element. Default is current "Internettabellen" header style}

\item{rowHeight}{height of the row(s)}

\item{...}{see create_style_generic}
}
\description{
create_header_style
}
