% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_index_df.R
\name{create_index_df}
\alias{create_index_df}
\title{create_index_df}
\usage{
create_index_df(data, pattern = "\\\\*\\\\*")
}
\arguments{
\item{data}{a data.frame}

\item{pattern}{a regular expression that should be looked for in the data}
}
\value{
a data.frame with row and col index of matching cells
}
\description{
Function returns a dataset with the indeces of those cells that are matching the selected pattern
}
