% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nested_header_style.R
\name{create_nested_header_style}
\alias{create_nested_header_style}
\title{create_nested_header_style}
\usage{
create_nested_header_style(
  wb,
  sheet,
  vars_level1,
  vars_level2,
  nesting = "even",
  startRow,
  ...
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{vars_level1}{vector of variable names (first level)}

\item{vars_level2}{vector of variable names (second level)}

\item{nesting}{if the variables should be evenly distributed, then provide the string "even" (= default). If the variables should NOT be evenly distributed then provide a numeric vector with the same length as vars_level1. The given integers indicate how many variables from level 2 are nested below one variable from level 1}

\item{startRow}{row in which the subheader should be placed (default = 2)}

\item{...}{see create_style_generic}
}
\description{
create_nested_header_style
}
