% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_style_df.R
\name{create_style_df}
\alias{create_style_df}
\title{create_style_df}
\usage{
create_style_df(wb)
}
\arguments{
\item{wb}{a workbook}
}
\value{
data.frame
}
\description{
Helper Function to delete NAs
}
