% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_style_generic.R
\name{create_style_generic}
\alias{create_style_generic}
\title{create_style_generic}
\usage{
create_style_generic(
  wb,
  sheet,
  ncol,
  text,
  startRow,
  style,
  vectorised = FALSE,
  style_only = FALSE,
  startCol = 1
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{ncol}{the number of columns}

\item{text}{a string or a vector}

\item{startRow}{row in which the content should be placed}

\item{style}{an openxlsx::Style element}

\item{vectorised}{if a vector is given for text, vectorised must be set to TRUE (default = FALSE)}

\item{style_only}{if set to TRUE, text will be ignored and only the style will be set (default = FALSE)}

\item{startCol}{col in which the function should start (default = 1)}
}
\description{
create_style_generic
}
