% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subheader_style.R
\name{create_subheader_style}
\alias{create_subheader_style}
\title{create_subheader_style}
\usage{
create_subheader_style(
  wb,
  sheet,
  ncol,
  text,
  startRow = 2,
  style = subheader_normal_10_left,
  ...
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{ncol}{the number of columns that the subheader should be long}

\item{text}{the text of the subheader}

\item{startRow}{row in which the subheader should be placed (default = 2)}

\item{style}{an openxlsx::Style element. Default is current "Internettabellen" subheader style}
}
\description{
create_subheader_style
}
