% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_style.R
\name{create_table_style}
\alias{create_table_style}
\title{create_table_style}
\usage{
create_table_style(
  wb,
  sheet,
  header,
  subheader,
  varnames,
  var_categories = NULL,
  data,
  datenquelle = NULL,
  footnotes = NULL,
  startRow = 1,
  gemeinde_format = FALSE,
  nesting = NULL,
  header_style = header_bold_12_left,
  subheader_style = subheader_normal_10_left,
  varnames_style = varname_normal_10_left,
  var_categories_style = varname_normal_10_left
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{header}{text for header}

\item{subheader}{text for subheader}

\item{varnames}{variable names}

\item{var_categories}{variable categories for nested header (nesting must be a numeric vector or 'even', otherwise var_categories will be ignored)}

\item{data}{a data.frame}

\item{datenquelle}{data source (default = NULL)}

\item{footnotes}{a character vector with footnotes. Please provide the correct order and do not forget the corresponding number (example: '1 Lorem Ipsum'). default = NULL}

\item{startRow}{row in which the header should be placed (default = 1)}

\item{gemeinde_format}{if TRUE formating for Gemeinde tables will be included (bold for Bezirk and Kanton) (default = FALSE)}

\item{nesting}{if the variables should be evenly distributed, then provide the string "even". If the variables should NOT be evenly distributed, please provide a numeric vector with the same length as var_categories The given integers indicate how many variables from level 2 are nested below one variable from level 1 (default=NULL)}

\item{header_style}{style for header (default = header_bold_12_left)}

\item{subheader_style}{style for subheader (default = subheader_normal_10_left )}

\item{varnames_style}{style for varnames (default = varname_normal_10_left )}

\item{var_categories_style}{style for var_categories (default = varname_normal_10_left )}
}
\description{
Function to create a complete table in excel in the format of the Internettabellen
}
