% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_reiter.R
\name{create_xlsx_reiter}
\alias{create_xlsx_reiter}
\title{Create a new Sheet (reiter) for data of the respective year}
\usage{
create_xlsx_reiter(xlsx_path, data, year, dataStart, dataStartCol = 2)
}
\arguments{
\item{xlsx_path}{path to the excel table where the row should be added}

\item{data}{data.frame for the respective year}

\item{year}{year (will later be the name of the sheet)}

\item{dataStart}{start of the data part of the sheet (below the header, subheader and varnames)}

\item{dataStartCol}{column index where the data starts. Default is `2`}
}
\value{
workvook object
}
\description{
Create a new Sheet (reiter) for data of the respective year
}
