% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_spalte.R
\name{create_xlsx_spalte}
\alias{create_xlsx_spalte}
\title{Append a new column for a new year}
\usage{
create_xlsx_spalte(
  xlsx_path,
  data,
  year,
  dataStart,
  limit = NULL,
  sheets = NULL
)
}
\arguments{
\item{xlsx_path}{path to the excel table where the row should be added}

\item{data}{data.frame for the respective year. The column that should be inserted must be named after the year (e.g. 2024). Can also be a list of data.frames, if multiple sheets should be added.}

\item{year}{year of the data}

\item{dataStart}{start of the data part of the sheet (below the header, subheader and varnames)}

\item{limit}{all values below this value will be replaced with `-`}

\item{sheets}{range of the sheets that should be edited. Default is NULL, e.g. first sheet will be edited.}
}
\value{
workbook object
}
\description{
Append a new column for a new year
}
