% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_zeile.R
\name{create_xlsx_zeile}
\alias{create_xlsx_zeile}
\title{Append a data row with the respective year}
\usage{
create_xlsx_zeile(xlsx_path, data, year, dataStart, sheets = NULL)
}
\arguments{
\item{xlsx_path}{path to the excel table where the row should be added}

\item{data}{the data.frame. Must contain one row with a column called "jahr" where the respective year is stored. Function wil only take the row where jahr==year. Can also be a list of data.frames if data should be added in multiple sheets}

\item{year}{the year of the data}

\item{dataStart}{start of the data part of the sheet (below the header, subheader and varnames)}

\item{sheets}{if a list of data.frames is given, the range of sheet indexes can be passed to the function here (e.g 1:3). Default is `NULL`}
}
\value{
a workbook object
}
\description{
Append a data row with the respective year
}
