% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_style.R
\name{customize_style}
\alias{customize_style}
\title{customize_style}
\usage{
customize_style(
  template = NULL,
  halign = NULL,
  valign = NULL,
  decoration = NULL,
  fontName = NULL,
  fontSize = NULL
)
}
\arguments{
\item{template}{can either be "header","subheader","varname", "data_year", "data_decimal" or an object of class openxlsx::Style. Default is Arial, 10, no decoration, no alignment, with thousands separator}

\item{halign}{horizontal alignment ("left", "right", "center" or "justify"). Default = NULL}

\item{valign}{vertical alignment ("top", "bottom" or "center"). Default = NULL}

\item{decoration}{font decoration ("bold", "strikeout", "italic", "underline", "underline2", "accounting" or"accounting2"). Default = NULL}

\item{fontName}{name of the Font. Default = "Arial"}

\item{fontSize}{font size in pt. Default = 10}
}
\value{
customized style of template openxlsx::Style
}
\description{
Function to customize a style for the common Internettabellen
}
