% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_styles_of_area.R
\name{get_styles_of_area}
\alias{get_styles_of_area}
\title{get_styles_of_area}
\usage{
get_styles_of_area(wb, sheet, rows, cols)
}
\arguments{
\item{wb}{a workbook}

\item{sheet}{sheet name}

\item{rows}{rows that should be looked at}

\item{cols}{columns that should be looked at}
}
\value{
list of styles including cols and rows of the style
}
\description{
Returns a list of styles present in a user defined area
}
