% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_superscript_numbers}
\alias{get_superscript_numbers}
\title{Get superscript numbers from text}
\usage{
get_superscript_numbers(text)
}
\arguments{
\item{text}{text}
}
\description{
Get superscript numbers from text
}
