% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_data_row.R
\name{insert_data_row}
\alias{insert_data_row}
\title{insert_data_row}
\usage{
insert_data_row(wb, sheet, data, insertRow, dataStartRow)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{data}{a data.frame with the data you want to insert}

\item{insertRow}{the index of the row the inserted data should start (styles will be taken from this particular row)}

\item{dataStartRow}{start of the data part of the sheet (below the header, subheader and varnames)}
}
\description{
Function to insert Data (similar to insert Row in Excel).
}
