% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rename_sheet}
\alias{rename_sheet}
\title{Rename sheet to given year}
\usage{
rename_sheet(wb, year)
}
\arguments{
\item{wb}{a workbook object}

\item{year}{the year of the data}
}
\description{
Rename sheet to given year
}
