% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_with_blanks.R
\name{replace_na_with_blanks}
\alias{replace_na_with_blanks}
\title{replace_na_with_blanks}
\usage{
replace_na_with_blanks(wb, style_df)
}
\arguments{
\item{wb}{a workbook}

\item{style_df}{result of create_style_df}
}
\description{
Replaces NAs with blanks in Workbook
}
