% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_nas_in_wb.R
\name{replace_nas_in_wb}
\alias{replace_nas_in_wb}
\title{replace_nas_in_wb}
\usage{
replace_nas_in_wb(wb)
}
\arguments{
\item{wb}{a workbook object}
}
\description{
Replaces NAs in Workbook so that they do not appear in the Header of a Table
}
