% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{replace_small_values}
\alias{replace_small_values}
\title{Replace small values for anonymisation}
\usage{
replace_small_values(
  wb,
  data,
  limit,
  col,
  replacement = "-",
  year,
  dataStart,
  sheet = 1
)
}
\arguments{
\item{wb}{a workbook object}

\item{data}{a data.frame with the data you want to insert}

\item{limit}{all values smaller than this will be repalced with the replacement}

\item{col}{index of the col}

\item{replacement}{replacement. Default is "-"}

\item{year}{year of the data}

\item{dataStart}{start of the data part of the sheet (below the header, subheader and varnames)}

\item{sheet}{sheet number}
}
\value{
workbook object
}
\description{
Replace small values for anonymisation
}
