% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tg_workbook.R
\name{save_tg_workbook}
\alias{save_tg_workbook}
\title{save_tg_workbook}
\usage{
save_tg_workbook(wb, filename, tg_header = TRUE, overwrite = FALSE)
}
\arguments{
\item{wb}{a workbook object}

\item{filename}{path where the file should be stored. Must have the extension .xlsx}

\item{tg_header}{if TRUE the header of the Internettabellen will be included (default = FALSE)}

\item{overwrite}{if TRUE, an existing file will be overwritten}
}
\description{
save_tg_workbook
}
