% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_numfmt.R
\name{special_numFmt}
\alias{special_numFmt}
\title{special_numfmt}
\usage{
special_numFmt(
  wb,
  sheet,
  index_df,
  dataStart,
  colShift = 0,
  below1000 = "#####0",
  from1000 = "###'##0",
  decimal = "###0.0",
  prefix = NULL,
  suffix = NULL
)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{name of the sheet}

\item{index_df}{result of `create_index_df`}

\item{dataStart}{index of the row where data starts in the excel file}

\item{colShift}{number of leading columns before the numeric data starts (difference in columns in excel and columns in data)}

\item{below1000}{numFmt for integer numbers below 1000 (default =  #####)}

\item{from1000}{numFmt for integer numbers from 1000 onwards (default = ###'##0)}

\item{decimal}{numFmt for decimal numbers (default = ###0.0)}

\item{prefix}{character string that will be put before the number. Every character value has to start with "\\". For Example if you wish to have to stars in front of the numer (e.g. `**100`), you need to write the suffix as follows. `\\*\\*`}

\item{suffix}{character string that will be put after the number. Same style as prefix}
}
\description{
Function applys specific numfmt on selected cells. Works only in cobination with `create_index_df`
}
