% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGexcel-package.R
\docType{package}
\name{TGexcel-package}
\alias{TGexcel}
\alias{TGexcel-package}
\title{TGexcel: TGexcel}
\description{
Das Package soll dabei helfen Internettabellen im jetzigen Format direkt aus R heraus erstellen zu können, ohne Excel verwenden zu müssen.
}
\seealso{
Useful links:
\itemize{
  \item \url{{{ URL }}}
  \item Report bugs at \url{{{ BugReports }}}
}

}
\author{
\strong{Maintainer}: Felix Lorenz \email{felix.lorenz@tg.ch}

}
\keyword{internal}
