% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_untertitel}
\alias{update_untertitel}
\title{Update subtitle to given year}
\usage{
update_untertitel(wb, year, sheet = 1)
}
\arguments{
\item{wb}{a workbook object}

\item{year}{the year of the data}

\item{sheet}{sheet number}
}
\description{
Update subtitle to given year
}
