% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_styles_to_wb.R
\name{write_styles_to_wb}
\alias{write_styles_to_wb}
\title{write_styles_to_wb}
\usage{
write_styles_to_wb(wb, sheet, filename)
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{filename}{where to store the resulting xlsx file}
}
\description{
Functions to see which cell styles are present in the workbook
}
