% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSuperSubScriptToCell.R
\name{addSuperSubScriptToCell}
\alias{addSuperSubScriptToCell}
\title{addSuperSubScriptToCell}
\usage{
addSuperSubScriptToCell(wb, sheet, row, col, texto, style, colour = "000000")
}
\arguments{
\item{wb}{a workbook object}

\item{sheet}{the name of the Sheet in the workbook object}

\item{row}{row index}

\item{col}{col index}

\item{texto}{data to write}

\item{style}{cell style}

\item{colour}{font colour in hex represenation (default is black -> default = '000000')}
}
\description{
See https://github.com/awalker89/openxlsx/issues/407#issuecomment-841870240
}
